bugs fix and code refactor.
