+++ src/pages/Home.tsx
@@ -1,26 +1,13 @@
-import { useState, useEffect } from "react";
-import {
-  fetchDepartments,
-  fetchFunctions,
-  fetchLocations,
-  fetchJobs,
-} from "../utils/api";
+import { useState } from "react";
 import Header from "../components/Header";
 import Joblist from "../components/Joblist";
-import type {
-  Job,
-  Department,
-  Function,
-  Location,
-  Filters,
-} from "../types/types";
 import Badges from "../components/Badges";
+import { useJobFilters } from "../hooks/useJobFilters";
+import type { Filters } from "../types/types";
+import { filterJobs } from "../utils/filterJobs";
 
 const Home = () => {
-  const [jobs, setJobs] = useState<Job[]>([]);
-  const [departments, setDepartments] = useState<Department[]>([]);
-  const [functions, setFunctions] = useState<Function[]>([]);
-  const [locations, setLocations] = useState<Location[]>([]);
+  const { jobs, departments, functions, locations } = useJobFilters();
   const [searchTerm, setSearchTerm] = useState<string>("");
   const [filters, setFilters] = useState<Filters>(() => {
     const stored = localStorage.getItem("filters");
@@ -28,23 +15,6 @@ const Home = () => {
       ? JSON.parse(stored)
       : { function: "", department: "", location: "" };
   });
-
-  useEffect(() => {
-    const fetchAll = async () => {
-      const [jobsData, departmentsData, functionsData, locationsData] =
-        await Promise.all([
-          fetchJobs(),
-          fetchDepartments(),
-          fetchFunctions(),
-          fetchLocations(),
-        ]);
-      setJobs(jobsData || []);
-      setDepartments(departmentsData || []);
-      setFunctions(functionsData || []);
-      setLocations(locationsData || []);
-    };
-    fetchAll();
-  }, []);
 
   const handleFilterChange = (key: string, value: string) => {
     const updated = { ...filters, [key]: value };
@@ -58,17 +28,7 @@ const Home = () => {
     localStorage.setItem("filters", JSON.stringify(cleared));
   };
 
-    const filteredJobs = jobs.filter((job) => {
-    const matchesFunction = !filters.function || job.function?.title === filters.function;
-    const matchesDepartment = !filters.department || job.department?.title === filters.department;
-    const matchesLocation = !filters.location || job.location?.title === filters.location;
-
-      const matchesSearch =
-    !searchTerm ||
-    job.title.toLowerCase().includes(searchTerm.toLowerCase());
-
-    return matchesFunction && matchesDepartment && matchesLocation && matchesSearch;
-  });
+  const filteredJobs = filterJobs(jobs, filters, searchTerm);
 
   return (
     <div>
